third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/live-edit-no-reveal-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests live edit feature.


Running: testLiveEditWithoutStepInWhenPausedDoesNotCauseCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Script execution paused.
Script execution resumed.
Cursor position is: (0, 0).

Running: testLiveEditWithStepInWhenPausedDoesNotCauseCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Did step into
Script execution resumed.
Script execution paused.
Script execution resumed.
Cursor position is: (0, 0).

Running: testLiveEditWithoutStepInWhenPausedThenStepIntoCausesCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Script execution paused.
Stepping into...
Did step into
Script execution resumed.
Script execution paused.
Script execution resumed.
Cursor position is: (8, 20).

Running: testLiveEditWithStepInWhenPausedThenStepIntoCausesCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Did step into
Stepping into...
Did step into
Script execution resumed.
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.
Cursor position is: (8, 4).
